[CHORE] Co-locate love helper with loveAnimal and route existing callers
The helper belongs next to the canonical rules it encodes — `loveAnimal`
itself — not in `applyAnimalFeedBuff` (where it landed initially just
because `isAnimalNeedingLove` happened to live there).

Moves `getNextLoveAvailableAt` (and its 4 tests) to `loveAnimal.ts` /
`loveAnimal.test.ts`. Two existing copies of the same gate math then
delegate to it:

  - `isAnimalNeedingLove` (applyAnimalFeedBuff.ts) → returns
    `getNextLoveAvailableAt(animal) < now`. `max(a, b) < now` is
    algebraically equivalent to `a < now && b < now`, so the boolean
    result is identical.
  - `isAnimalNeedsLove` private guard (animalMachine.ts) → same shape.

`loveAnimal` itself keeps its two inline throws unchanged — each
throws a distinct user-facing error ("not been sleeping for more than
8 hours" vs "loved in the last 8 hours") that a single max-check
would collapse, so consolidating there would be a behavior change.
View-layer copies in `Barn.tsx` / `HenHouse.tsx` are left for a
follow-up to keep this PR's review surface focused.
