+++ src/features/game/events/landExpansion/loveAnimal.ts
@@ -5,8 +5,23 @@ import {
 } from "features/game/lib/animals";
 import { isCollectibleBuilt } from "features/game/lib/collectibleBuilt";
 import { AnimalLevel, ANIMALS, AnimalType } from "features/game/types/animals";
-import { GameState, LoveAnimalItem } from "features/game/types/game";
+import { Animal, GameState, LoveAnimalItem } from "features/game/types/game";
 import { produce } from "immer";
+
+/**
+ * Earliest timestamp at which this animal's current sleep cycle would
+ * permit `loveAnimal`. Mirrors the two throw-gates below: both
+ * `asleepAt + period` and `lovedAt + period` must have elapsed, where
+ * `period = (awakeAt - asleepAt) / 3`. Callers should still check
+ * `t < animal.awakeAt` — once the animal is awake the cycle is over
+ * and no further love applies. The returned value can therefore be
+ * `>= awakeAt` when no slot remains in this cycle (e.g. both slots
+ * already used); the caller decides how to render that.
+ */
+export function getNextLoveAvailableAt(animal: Animal): number {
+  const third = (animal.awakeAt - animal.asleepAt) / 3;
+  return Math.max(animal.asleepAt + third, animal.lovedAt + third);
+}
 
 export type LoveAnimalAction = {
   type: "animal.loved";
